# Add tag support in statsd reporter | Upgrade go-statsd-client