--- a/glide.yaml
+++ b/glide.yaml
@@ -1,7 +1,7 @@
 package: github.com/uber-go/tally
 import:
 - package: github.com/cactus/go-statsd-client
-  version: ~3.1.0
+  version: ^5.0.0
   subpackages:
   - statsd
 - package: github.com/m3db/prometheus_client_golang