--- a/statsd/example/statsd_main.go
+++ b/statsd/example/statsd_main.go
@@ -25,17 +25,22 @@ import (
 	"math/rand"
 	"time"
 
+	"github.com/cactus/go-statsd-client/statsd"
 	"github.com/uber-go/tally"
 	statsdreporter "github.com/uber-go/tally/statsd"
-
-	"github.com/cactus/go-statsd-client/statsd"
 )
 
 // To view statsd emitted metrics locally you can use
 // netcat with "nc 8125 -l -u"
 func main() {
-	statter, err := statsd.NewBufferedClient("127.0.0.1:8125",
-		"stats", 100*time.Millisecond, 1440)
+	statsdClientCfg := statsd.ClientConfig{
+		Address:       "127.0.0.1:8125",
+		Prefix:        "stats",
+		FlushInterval: 100 * time.Millisecond,
+		FlushBytes:    1440,
+		TagFormat:     statsd.InfixComma,
+	}
+	statter, err := statsd.NewClientWithConfig(&statsdClientCfg)
 	if err != nil {
 		log.Fatalf("could not create statsd client: %v", err)
 	}
@@ -50,7 +55,10 @@ func main() {
 	}, 1*time.Second)
 	defer closer.Close()
 
-	counter := scope.Counter("test-counter")
+	counter := scope.Tagged(map[string]string{
+		"host": "local",
+		"env":  "dev",
+	}).Counter("test-counter")
 
 	gauge := scope.Gauge("test-gauge")